requirements.txt
apache-airflow>=2.7.0
mysql-connector-python>=8.0.0